requirements.txt
psycopg2-binary~=2.9.9
pydantic~=2.9.2
aiogram~=3.13.1
python-dotenv~=1.0.1

